# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EduFlow LMS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /playlists:
    get:
      operationId: listPlaylists
      tags: [playlists]
      summary: List all active playlists
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Playlists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlaylistSummary"
    post:
      operationId: createPlaylist
      tags: [playlists]
      summary: Create a playlist (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playlist"
  /playlists/{id}:
    get:
      operationId: getPlaylist
      tags: [playlists]
      summary: Get a single playlist with its videos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Playlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaylistDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePlaylist
      tags: [playlists]
      summary: Update a playlist (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playlist"
    delete:
      operationId: deletePlaylist
      tags: [playlists]
      summary: Delete a playlist (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /content:
    get:
      operationId: listContent
      tags: [content]
      summary: List all content
      parameters:
        - name: playlistId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Content list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentItem"
    post:
      operationId: createContent
      tags: [content]
      summary: Create content (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentItem"
  /content/{id}:
    get:
      operationId: getContent
      tags: [content]
      summary: Get a single content item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Content item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateContent
      tags: [content]
      summary: Update content (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentItem"
    delete:
      operationId: deleteContent
      tags: [content]
      summary: Delete content (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Comments ─────────────────────────────────────────────────────────────────
...
  /content/{id}/like:
    get:
      operationId: getLikeStatus
      tags: [likes]
      summary: Get like status and count for a video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Like status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeStatus"
    post:
      operationId: toggleLike
      tags: [likes]
      summary: Toggle like on a video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeStatus"

  # ── Bookmarks ────────────────────────────────────────────────────────────────
...
  /tutors/{id}:
    get:
      operationId: getTutor
      tags: [tutors]
      summary: Get a tutor with their playlists
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tutor detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TutorDetail"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # ── Auth ─────────────────────────────────────────────────────────────────
...
    PlaylistInput:
      type: object
      required: [title, description]
      properties:
        title:
          type: string
        description:
          type: string
        image:
          type: string
        status:
          type: string
          enum: [active, inactive]
    PlaylistUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        image:
          type: string
        status:
          type: string
          enum: [active, inactive]
    Playlist:
      type: object
      required: [id, title, description, status]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        image:
          type: ["string", "null"]
        status:
          type: string
        tutorId:
          type: integer
        createdAt:
          type: string
          format: date-time
    PlaylistSummary:
      type: object
      required: [id, title, description, status]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        image:
          type: ["string", "null"]
        status:
          type: string
        tutorId:
          type: integer
        tutorName:
          type: ["string", "null"]
        videoCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    PlaylistDetail:
      type: object
      required: [id, title, description, status, videos]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        image:
          type: ["string", "null"]
        status:
          type: string
        tutorId:
          type: integer
        tutorName:
          type: ["string", "null"]
        tutorProfession:
          type: ["string", "null"]
        videoCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        videos:
          type: array
          items:
            $ref: "#/components/schemas/ContentItem"

    # ── Content ───────────────────────────────────────────────────────────────
    ContentInput:
      type: object
      required: [playlistId, title, description, videoUrl]
      properties:
        playlistId:
          type: integer
        title:
          type: string
        description:
          type: string
        videoUrl:
          type: string
        status:
          type: string
          enum: [active, inactive]
    ContentUpdate:
      type: object
      properties:
        playlistId:
          type: integer
        title:
          type: string
        description:
          type: string
        videoUrl:
          type: string
        status:
          type: string
          enum: [active, inactive]
    ContentItem:
      type: object
      required: [id, title, description, videoUrl, status]
      properties:
        id:
          type: integer
        playlistId:
          type: integer
        tutorId:
          type: integer
        title:
          type: string
        description:
          type: string
        videoUrl:
          type: string
        status:
          type: string
        likeCount:
          type: integer
        commentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    ContentDetail:
      type: object
      required: [id, title, description, videoUrl, status]
      properties:
        id:
          type: integer
        playlistId:
          type: integer
        playlistTitle:
          type: ["string", "null"]
        tutorId:
          type: integer
        tutorName:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: string
        videoUrl:
          type: string
        status:
          type: string
        likeCount:
          type: integer
        commentCount:
          type: integer
        createdAt:
          type: string
          format: date-time

    # ── Comments ──────────────────────────────────────────────────────────────
...
    LikeStatus:
      type: object
      required: [liked, count]
      properties:
        liked:
          type: boolean
        count:
          type: integer

    # ── Bookmarks ─────────────────────────────────────────────────────────────
...
    TutorDetail:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        profession:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        playlistCount:
          type: integer
        playlists:
          type: array
          items:
            $ref: "#/components/schemas/PlaylistSummary"