# lib/api-zod/src/generated/types/playlistDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EduFlow LMS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContentItem } from './contentItem';

export interface PlaylistDetail {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  image?: string | null;
  status: string;
  tutorId?: number;
  /** @nullable */
  tutorName?: string | null;
  /** @nullable */
  tutorProfession?: string | null;
  videoCount?: number;
  createdAt?: Date;
  videos: ContentItem[];
}