# lib/api-zod/src/generated/types/playlistInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EduFlow LMS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlaylistInputStatus } from './playlistInputStatus';

export interface PlaylistInput {
  title: string;
  description: string;
  image?: string;
  status?: PlaylistInputStatus;
}